pyproject.toml
[tool.poetry]
name = "protalkapi"
version = "0.1.0"
description = ""
authors = ["Кирил Бессонов <ki48bes2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.40"
pydantic = "^2.11.3"
werkzeug = "^3.1.3"
uvicorn = "^0.34.2"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
email-validator = "^2.2.0"
pydantic-settings = "^2.9.1"
python-multipart = "^0.0.20"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
